Fix suprious trigger of autofillers
On Windows AltGr+y is translated to Ctrl+Alt+y. Since the autofiller is
bound to Ctrl+y and that we are not checking that other modifiers are
disabled, AltGr+y (which maps to useful characters on some keymaps)
incorrectly triggers the autofiller.

Fixed by not triggering when other modifiers are up.